# openapi.json
{
  "components": {
    "schemas": {
...
      "Body_createReview": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "The permit set, as a PDF.",
            "title": "File",
            "type": "string"
          },
          "review_options": {
            "default": "{}",
            "description": "A JSON-encoded ReviewOptions object.",
            "title": "Review Options",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_createReview",
        "type": "object"
      },
      "ConfigResponse": {
        "properties": {
          "defaults": {
            "$ref": "#/components/schemas/ReviewOptions"
          },
          "editions": {
            "items": {
              "$ref": "#/components/schemas/Edition"
            },
            "title": "Editions",
            "type": "array"
          },
          "mail": {
            "$ref": "#/components/schemas/MailStatus"
          },
          "max_pages": {
            "title": "Max Pages",
            "type": "integer"
          },
          "max_upload_mb": {
            "title": "Max Upload Mb",
            "type": "integer"
          },
          "occupancy_groups": {
            "items": {
              "$ref": "#/components/schemas/OccupancyGroup"
            },
            "title": "Occupancy Groups",
            "type": "array"
          },
          "retain_days": {
            "title": "Retain Days",
            "type": "integer"
          },
          "rules": {
            "description": "Rule ids registered in this build.",
            "items": {
              "type": "string"
            },
            "title": "Rules",
            "type": "array"
          },
          "severities": {
            "description": "Severity floor choices, most severe first.",
            "items": {
              "enum": [
                "CRITICAL",
                "HIGH",
                "MEDIUM",
                "LOW"
              ],
              "type": "string"
            },
            "title": "Severities",
            "type": "array"
          },
          "stages": {
            "description": "Ordered stage labels a running job moves through.",
            "items": {
              "type": "string"
            },
            "title": "Stages",
            "type": "array"
          }
        },
        "required": [
          "occupancy_groups",
          "editions",
          "severities",
          "stages",
          "max_upload_mb",
          "max_pages",
          "retain_days",
          "rules",
          "mail",
          "defaults"
        ],
        "title": "ConfigResponse",
        "type": "object"
      },
      "Downloads": {
        "description": "V4 signed URLs, fetched straight from Cloud Storage by the browser.",
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "findings_json": {
            "title": "Findings Json",
            "type": "string"
          },
          "markup_pdf": {
            "title": "Markup Pdf",
            "type": "string"
          }
        },
        "required": [
          "markup_pdf",
          "findings_json",
          "expires_at"
        ],
        "title": "Downloads",
        "type": "object"
      },
      "Edition": {
        "properties": {
          "available": {
            "description": "False editions are listed and disabled, not hidden.",
            "title": "Available",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "available"
        ],
        "title": "Edition",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "Stable machine-readable code, e.g. `rate_limited`.",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "description": "Prose safe to show a person.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "The only shape any failure takes.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Job": {
        "properties": {
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "downloads": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Downloads"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present only while state is `done`."
          },
          "elapsed_seconds": {
            "title": "Elapsed Seconds",
            "type": "number"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "filename": {
            "description": "Basename as uploaded. Never a path.",
            "title": "Filename",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/ReviewOptions"
          },
          "pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceProfile"
              },
              {
                "type": "null"
              }
            ],
            "description": "What kind of PDF was uploaded, measured at admission."
          },
          "stage": {
            "description": "Index into `stages`.",
            "minimum": 0.0,
            "title": "Stage",
            "type": "integer"
          },
          "stage_label": {
            "title": "Stage Label",
            "type": "string"
          },
          "stages": {
            "items": {
              "type": "string"
            },
            "title": "Stages",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "enum": [
              "queued",
              "running",
              "done",
              "error"
            ],
            "title": "State",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Summary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "filename",
          "state",
          "stage",
          "stage_label",
          "stages",
          "options",
          "bytes",
          "created_at",
          "elapsed_seconds"
        ],
        "title": "Job",
        "type": "object"
      },
      "MailStatus": {
        "properties": {
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "configured",
          "status"
        ],
        "title": "MailStatus",
        "type": "object"
      },
      "OccupancyGroup": {
        "properties": {
          "id": {
            "description": "Value submitted with a review, e.g. `A-3`.",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label"
        ],
        "title": "OccupancyGroup",
        "type": "object"
      },
      "ReviewAccepted": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "ReviewAccepted",
        "type": "object"
      },
      "ReviewOptions": {
        "additionalProperties": false,
        "description": "The review parameters. Mirrors `fbcreview.options.ReviewOptions`.",
        "properties": {
          "edition": {
            "default": "fbc2023",
            "title": "Edition",
            "type": "string"
          },
          "email_to": {
            "description": "Inert unless SMTP is configured on the server. Not surfaced in the web client.",
            "items": {
              "type": "string"
            },
            "title": "Email To",
            "type": "array"
          },
          "include_measured": {
            "default": true,
            "title": "Include Measured",
            "type": "boolean"
          },
          "include_verified": {
            "default": true,
            "title": "Include Verified",
            "type": "boolean"
          },
          "min_severity": {
            "default": "LOW",
            "enum": [
              "CRITICAL",
              "HIGH",
              "MEDIUM",
              "LOW"
            ],
            "title": "Min Severity",
            "type": "string"
          },
          "notes": {
            "default": "",
            "maxLength": 2000,
            "title": "Notes",
            "type": "string"
          },
          "occupancy_group": {
            "default": "A-3",
            "title": "Occupancy Group",
            "type": "string"
          },
          "project_name": {
            "default": "",
            "maxLength": 200,
            "title": "Project Name",
            "type": "string"
          },
          "sprinklered": {
            "default": true,
            "title": "Sprinklered",
            "type": "boolean"
          }
        },
        "title": "ReviewOptions",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Upload a vector permit set, choose the review parameters, download a marked-up PDF and a findings register. No model calls in the request path.",
    "summary": "Deterministic Florida Building Code plan review.",
    "title": "FBC Code Review",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/config": {
      "get": {
        "operationId": "getConfig",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Config",
        "tags": [
          "config"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "operationId": "getJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/review": {
      "post": {
        "operationId": "createReview",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_createReview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Accept a permit set and start a review.",
        "tags": [
          "reviews"
        ]
      }
    },
...
  }
}